02.코드 최적화하기

+++ src/main/java/com/fastcampus/sns/controller/PostController.java
@@ -57,7 +57,7 @@ public class  PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId, Authentication authentication) {
+    public Response<Long> likeCount(@PathVariable Integer postId, Authentication authentication) {
         return Response.success(postService.likeCount(postId));
     }
